refactor(contracts): share the jobs-query semantics between API and demo
The jobs-list filter, sort, compensation sort values, and pagination
envelope existed as two hand-copied TypeScript twins - one in the read
model, one in the demo adapter - kept aligned only by hand-re-encoded test
expectations. A filter fix in one could silently miss the other.

- new packages/contracts/src/jobs-query.ts owns the shared semantics:
  filterJob (the non-SQL filter core), compareJobs with a normalizeSqlText
  option (the demo's emulation of SQLite ORDER BY collation), the
  compensation sort values, STATE_RANK, compareValues, timestamps, paginate
- the read model consumes the shared module (its deleted-facet filtering
  stays in SQL WHERE, unchanged); STATE_RANK still feeds the SQL rank CASE
- the demo adapter keeps only a thin filterJob wrapper for the deleted
  facet it must emulate client-side, and passes normalizeSqlText per sort
  field as before
- both twins' remaining local comparators (artifacts, activity, workflow
  runs) now share the one compareValues

Demo 'matches production' tests keep passing - now by construction rather
than by parallel maintenance.
