+++ packages/contracts/src/index.ts
@@ -1,6 +1,7 @@
 export * from "./schemas.js";
 export * from "./rpc.js";
 export * from "./demo-fixture-contract.js";
+export * from "./jobs-query.js";
 
 // Operations / Read-Side projection types — re-exported from
 // @jobctrl/domain-types so apps/api (which depends on contracts but
